chore: relocate copy-button component & remove unused kanban component
